Problemset02: Extend snakes and Ladders.
The full task is in problemset02/p2unibe-2010-week02.mdown. Summary:
Modify the Snakes and Ladders game from the lecture in the following ways: Add two new types of squares: LargeSquare and TrapDoor, Write a unit test LargeSquareTest that verifies that LargeSquare works as intended. Add more test methods to TrapDoorTest that verify that your TrapDoor works as intended. Modify the game such that the Game.main method will run on a field with a large squares and a trap door. Add API documentation to the snakes.ISquare interface according to the Oracle-provided rules.