DR Fix problemset03 for next week!
What's the scope of these variables?
I need your UML and your contracts (see Oscar's slieds) to finish my corrections
Also something is very wrong if you run ShowImage and then
	- translate 100 50
	- scale 0.5
	- rotate 1.5

And your code is very noisy if I enter a wrong transformation one of the classes has to check it and
it should not throw an AssertionError it should throw an WrongStringException or something like that.