Accepted! You passed problemset 5.2!
Everything looks ok:
all the different phases are in your code and you made a good decision for your genecards (a separate class ^^)
It would be nice to interact (really play the game) instead of just watching but thats another point
But you should at least show if a player buys a gene or decides to move instead of drift!
Also I realy like your test coverage it's just awesome!
Keep up the good work, can't wait for a GUI (and some sketches)