Solved problemset08.
We drawed some ideas to construct a GUI and shortly explained which
seems to be the best one.
Further there's a prototype to show how our favorite GUI works.
We implemented a GUI in our Ursuppe game that simply shows what our
Game#toString-Method composes and added an Output-Class to organise
where the game is printed (so that now our tests run silently).

	new file:   problemset05/src/UrsuppeGUI/Ursuppe.java
	new file:   problemset05/src/ursuppe/ConsoleOutput.java
	new file:   problemset05/src/ursuppe/GUIOutput.java
	modified:   problemset05/src/ursuppe/Game.java
	modified:   problemset05/src/ursuppe/GameTest.java
	modified:   problemset05/src/ursuppe/GenePool.java
	new file:   problemset05/src/ursuppe/IntelligenceGene.java
	new file:   problemset05/src/ursuppe/NullOutput.java
	new file:   problemset05/src/ursuppe/Output.java
	new file:   problemset08/GUI_1.JPG
	new file:   problemset08/GUI_2.JPG
	new file:   problemset08/GUI_3.JPG
	new file:   problemset08/GUI_4.JPG
	new file:   problemset08/GUI_5.JPG
	new file:   problemset08/OurFavoriteGUI.mdown
	new file:   problemset08/paper_prototype.JPG
	new file:   problemset09/problemset 9.html